+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Remember accepted/ingested bytes for log outcomes. ([#4886](https://github.com/getsentry/relay/pull/4886))
 - Add `gen_ai.response.tokens_per_second` span attribute on AI spans. ([#4883](https://github.com/getsentry/relay/pull/4883))
 - Add support for playstation data requests. ([#4870](https://github.com/getsentry/relay/pull/4870))
+- Add snapshot testing for python integration tests. ([#4877](https://github.com/getsentry/relay/pull/4877))
 
 ## 25.6.2